+++ changelog.md
@@ -8,6 +8,7 @@ Codion Change Log
 ### is.codion.swing.common.ui
 - FilterTable now automatically creates columns, Builder.columns() now allows column builder configuration.
 - FilterTableColumn can no longer be instantiated outside FilterTable, Builder.build() removed, related refactoring.
+- CalendarPanel bug fixed, clicking day filler labels now sets the year as well as the month.
 ### is.codion.framework.domain
 - Entities.primaryKeys() overloaded with Collection.
 ### is.codion.framework.model